requirements/base.txt
alembic==0.9.9
amqp==2.2.2
backcall==0.1.0
billiard==3.5.0.3
celery==4.1.0
click==6.7
decorator==4.3.0
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Migrate==2.1.1
flask-shell-ipython==0.3.0
Flask-SQLAlchemy==2.3.2
ipython==6.3.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.7
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
flask-wtf
pytest-flask
requests
